🌿 [codebase-cleanup] client(module_core): consolidate test helpers [step 12/45]
Seven mocks were re-declared across 13 test files instead of coming from
test/helpers/test_helpers.dart: MockPermissionRepository and
MockNotificationCanceller (6 files each), MockRelayHttpApiClient (5),
MockAuthSession and MockSessionDetailLoadService (4), MockFailureReporter (3),
MockRoomKeyStorage and FakeAuthSession (2). Every Mock duplicate was a
byte-identical one-liner and both FakeAuthSession copies hashed identically, so
they move to the helper unchanged.

One real divergence: session_viewing_service_test declared its own
FakeLifecycleSource whose emit method was named emit rather than the helper's
emitState. Same behaviour; the call sites now use the helper's name.

Verification: dart analyze --fatal-infos clean and 1,172 tests pass in
client/module_core.
