+++ client/module_core/test/services/notification_preferences_service_test.dart
@@ -5,8 +5,7 @@ import "package:rxdart/rxdart.dart";
 import "package:sesori_dart_core/sesori_dart_core.dart";
 import "package:sesori_shared/sesori_shared.dart";
 import "package:test/test.dart";
-
-class MockAuthSession() extends Mock implements AuthSession;
+import "../helpers/test_helpers.dart";
 
 class MockNotificationPreferencesRepository() extends Mock implements NotificationPreferencesRepository;
